software/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.20.0
pyyaml>=6.0
adafruit-circuitpython-ads1x15>=2.2.0
RPi.GPIO>=0.7.0
adafruit-blinka>=8.0.0

